Get the correct fontsize in testcase 6.

+++ src/main/java/edworld/pdfreader4humans/impl/MainPDFComponentLocator.java
@@ -228,7 +228,8 @@ public class MainPDFComponentLocator implements PDFComponentLocator {
 			}
 
 			private float fontsz(TextPosition textPosition) {
-				return textPosition.getFontSize() == 1F ? textPosition.getFontSizeInPt() : textPosition.getFontSize();
+				return textPosition.getFontSize() == 1F ? textPosition.getFontSizeInPt()
+						: (int) textPosition.getFontSize();
 			}
 
 			private Component findOverlappingHorizontalShape(TextPosition textPosition) {
